- completed setup account flow - moved search business to staff folder
